requirements.txt
pytorch
gdown
numpy
opencv==4.1.1
matplotlib
